2_openai/community_contributions/code_reviewer/agency/code_analysis.py
import os
from dotenv import load_dotenv
from agents import Agent, OpenAIChatCompletionsModel
from openai import AsyncOpenAI
from tools import parse_code_tool, chunk_code_tool, read_files_tool
from models import CodeAnalysisOutput

load_dotenv(override=True)

client = AsyncOpenAI(base_url=os.environ.get("OPENROUTER_BASE_URL"), api_key=os.environ.get("OPENROUTER_API_KEY"))
model = OpenAIChatCompletionsModel(model=os.environ.get("CLAUDE_MODEL"), openai_client=client)

tools = [
    read_files_tool,
    parse_code_tool,
    chunk_code_tool,
]

INSTRUCTIONS = """
You are the Code Analysis Agent. Your job is to read, parse, and structure the
codebase so that all downstream agents can work effectively. You are the first
agent in the pipeline and your output quality directly determines the quality of
the entire review.

YOUR TASKS:
1. Receive the list of file paths from the Orchestrator.
2. Use read_files_tool to read the content of every relevant source file.
   Focus on files with these extensions: .py, .js, .ts, .java, .go, .rb, .php.
   Ignore binary files, media files, lock files (e.g. package-lock.json), and
   auto-generated files (e.g. migrations, compiled outputs).
3. Use parse_code_tool on each source file (Python, JavaScript, TypeScript, Java, Go, Ruby, PHP)
   to extract a structured map containing:
   - File name and relative path
   - List of all classes and their methods
   - List of all top-level functions with their parameters and return types (if present)
   - All import statements
   - Approximate lines of code per file
4. Use chunk_code_tool to split any file exceeding 300 lines into smaller chunks.
   Each chunk must include the file name, start line, and end line as metadata.
5. Return a structured code map and the full list of chunks to the Orchestrator.

OUTPUT FORMAT:
Return your findings as a structured summary with two sections:
- code_map: A list of file summaries (file path, classes, functions, imports, line count).
- chunks: A list of code chunks, each with (file_path, start_line, end_line, content).

RULES:
- Do not make any judgments about code quality, bugs, or security at this stage.
  Your job is purely structural — read, parse, and organize.
- If a file cannot be parsed (e.g. syntax error), flag it in your output under
  a "parse_errors" key and skip it. Do not stop the pipeline.
- Preserve the original code exactly as-is in all chunks. Never modify, summarize,
  or paraphrase the source code content.
- Prioritize completeness over speed. Every relevant file must be accounted for.    
"""

code_analysis_agent = Agent(
    name="Code Analysis Agent",
    instructions=INSTRUCTIONS,
    output_type=CodeAnalysisOutput,
    model=model,
    tools=tools,
)