2_openai/community_contributions/code_reviewer/requirements.txt
gradio
python-dotenv
gitpython
pydantic
openai-agents
tree-sitter 
tree-sitter-languages